Add unittest to document existing behaviour of qasm2 rz and crz gates in pyzx.
The OpenQASM 2 spec defines the following:
```
gate rz(phi) a { u1(phi) a; }
gate crz(lambda) a,b
{
    u1(lambda/2) b;
    cx a,b;
    u1(-lambda/2) b;
    cx a,b;
}
gate cu1(lambda) a,b
{
    u1(lambda/2) a;
    cx a,b;
    u1(-lambda/2) b;
    cx a,b;
    u1(lambda/2) b;
}
```

pyzx's `ZPhase` gate and `CRZ` gates are equivalent to, respectively, OpenQASM 2's `rz` (a synonym for `u1`) and `crz` gates. Confusingly, the `crz` gate is not simply a controlled version of the `rz` gate; that gate is named `cu1`.

Meanwhile, the OpenQASM 3 spec defines:
```
gate rz(λ) a { gphase(-λ/2); U(0, 0, λ) a; }
gate crz(θ) a, b { ctrl @ rz(θ) a, b; }
```

This has the advantage that `crz` is now simply a controlled version of `rz`, at the expense of changing the meaning of `rz` so that it no longer matches `u1`/`ZPhase`.

This difference in conventions leads to user confusion and coding errors, as seen in issues #102 and #103. As well, changing the behaviour of `qasmparser` depending on the version of OpenQASM used is a blocker for issue #116.