Document qasm gate semantics when imported into pyzx, with a unit test.
There are number of differences between pyzx's and qiskit's interpretation of qasm gates, leading to user confusion. This PR adds a unit test to document these differences. Users or developers may trigger a test failure to see an error message showing the difference between the pyzx and qiskit versions of a qasm gate.