In `full_reduce`, raise error if input graph contains `H_BOX`.
The function requires that the input is a ZX-diagram (i.e., no `H_BOX`es). See #161 and #200 for context.

Note: This does not guard `clifford_simp` or other simplification subroutines, which will still silently fail if there are `H_BOX`es. Putting a check into each subroutine would be redundant and expensive.