Allow random seeds
Allow (optionally) user-specified seeds for random graph generation. This is helpful (e.g. in benchmarking datasets) as it allows one to regenerate the same graph just by specifying the same seed (which avoids having to save the graph, e.g. as json, if it will be needed again at a later date).